fix lightbox
