+++ src/Utils/localImagesHelpers.ts
@@ -1,5 +1,6 @@
 import { LOCAL_PROFILE_IMAGE_KEY } from "Apps/Settings/Routes/EditProfile/Components/SettingsEditProfileImage/utils/constants"
 import localforage from "localforage"
+import { useCallback, useEffect, useState } from "react"
 
 // Expiritation time is 5 minutes
 // TODO: Decrease number
@@ -146,7 +147,7 @@ const prepareImage = (image: LocalImage, expires: string) => {
     data: image.data,
     height: image.height,
     width: image.width,
-    aspectRatio: image.width / image.height,
+    aspectRatio: image.width / (image.height || 1),
   }
 
   return JSON.stringify(imageToStore)
@@ -156,3 +157,36 @@ export const isImageVersionAvailable = (
   image: { readonly versions: any } | null,
   version: string
 ) => !!image?.versions?.includes(version)
+
+export const useLocalImage = (image: any) => {
+  const [localImage, setLocalImage] = useState<LocalImage | null>(null)
+
+  const fetchLocalImage = useCallback(() => getLocalImage(image.internalID!), [
+    image.internalID,
+  ])
+
+  const isImageAvailable = isImageVersionAvailable(image, "large")
+
+  const changeLocalImage = async () => {
+    if (isImageAvailable || !image.internalID) {
+      setLocalImage(null)
+
+      return
+    }
+
+    try {
+      const image = await fetchLocalImage()
+
+      setLocalImage(image)
+    } catch (error) {
+      console.log(error)
+    }
+  }
+
+  useEffect(() => {
+    changeLocalImage()
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [image.internalID])
+
+  return localImage
+}
